Change the Language test to follow conventions
